Optimized copy-to-clipboard
