+++ src/scripts/copyToClipboard/copyToClipboard.module.css
@@ -1,0 +1,27 @@
+.copyCodeButtonWrapper {
+  padding: var(--space-2x) var(--space-3x) var(--space-0x) var(--space-0x);
+  margin-top: 0 !important;
+  z-index: 2;
+  position: absolute;
+  transform: translateY(2px);
+  right: 0;
+}
+
+.copyCodeButtonWrapper > button {
+  padding: var(--space-1x);
+  display: flex;
+}
+
+.copyCodeButtonWrapper > button > img {
+  width: 16px;
+  height: 16px;
+}
+
+pre.copyCodeButtonPlaceholderHidden::before {
+  display: none;
+}
+
+.codeWrapper {
+  position: relative;
+  margin-top: var(--space-2x);
+}
