--- src/styles/copy-to-clipboard.css
@@ -1,51 +1,0 @@
-/* TODO: Don't let button overlap code (esp. on mobile) */
-/* TODO: Inline these styles to components */
-.copy-code-button-wrapper {
-  padding: var(--space-2x) var(--space-3x) var(--space-0x) var(--space-0x);
-  margin-top: 0 !important;
-  z-index: 2;
-  position: absolute;
-  transform: translateY(2px);
-  right: 0;
-}
-
-.copy-code-button-wrapper > button {
-  padding: var(--space-1x);
-  display: flex;
-}
-
-.copy-code-button-wrapper > button > img {
-  width: 16px;
-  height: 16px;
-}
-
-pre::before {
-  content: "";
-  width: 28px;
-  height: 28px;
-  margin: var(--space-2x) var(--space-2x) var(--space-0x) var(--space-0x);
-  z-index: 2;
-  position: absolute;
-  transform: translateY(5px);
-  top: 0;
-  right: 0;
-  border: 2px;
-  border-style: solid;
-  border-color: var(--blue-600);
-  border-radius: var(--border-radius-primary);
-  border-width: var(--border-width-secondary);
-  background-color: white;
-  background-image: url("/assets/icons/copyIcon.svg");
-  background-repeat: no-repeat;
-  background-size: 16px 16px;
-  background-position: center;
-}
-
-pre.copy-code-button-placeholder-hidden::before {
-  display: none;
-}
-
-.code-wrapper {
-  position: relative;
-  margin-top: var(--space-2x);
-}
